+++ lib/models/person.dart
@@ -1,3 +1,4 @@
+import 'package:gezamycar/enums/auth-result-status.dart';
 import 'package:gezamycar/exceptions/my_exception.dart';
 import 'package:gezamycar/utils/constants.dart';
 
@@ -6,7 +7,6 @@ abstract class Person {
   String _fistName;
   String _lastName;
   String _gender;
-  String _password;
 
   void id(String uid) => _uid = uid;
 
@@ -18,12 +18,6 @@ abstract class Person {
       ? throw MyException(kFieldIsRequired)
       : _lastName = lastName;
 
-  void password(String password) => password.isEmpty
-      ? throw MyException(kFieldIsRequired)
-      : !_isValidPassword(password)
-          ? throw MyException(kPasswordShort)
-          : _password = password;
-
   void gender(String gender) => _gender = gender;
 
   String getUid() => _uid;
@@ -34,17 +28,13 @@ abstract class Person {
 
   String getGender() => _gender;
 
-  String getPassword() => _password;
-
   void login();
 
-  Future<int> signUp();
+  Future<AuthResultStatus> signUp();
 
   void resetPassword();
 
-  bool _isValidPassword(String password) => password.length > 6;
-
   @override
   String toString() =>
-      'FirstName: $_fistName, LastName: $_lastName, Gender: $_gender';
+      'Id: $_uid, FirstName: $_fistName, LastName: $_lastName, Gender: $_gender';
 }