+++ lib/services/auth_services.dart
@@ -1,8 +1,11 @@
 import 'package:firebase_auth/firebase_auth.dart';
+import 'package:gezamycar/enums/auth-result-status.dart';
+import 'package:gezamycar/exceptions/auth-exception-handler.dart';
 import 'package:gezamycar/exceptions/my_exception.dart';
 
 class AuthServices {
   final _auth = FirebaseAuth.instance;
+  AuthResultStatus _status;
 
   // signIn user
   Future<User> signIn(String emailAddress, String password) async {
@@ -16,14 +19,18 @@ class AuthServices {
   }
 
   // create user account
-  Future<User> singUp(String emailAddress, String password) async {
+  Future<AuthResultStatus> singUp(String emailAddress, String password) async {
     try {
       UserCredential result = await _auth.createUserWithEmailAndPassword(
           email: emailAddress, password: password);
-      return result.user;
+      if (result.user != null) {
+        _status = AuthResultStatus.successful;
+      }
     } catch (e) {
-      throw MyException(e.toString());
+      _status = AuthExceptionHandler.handleException(e);
     }
+
+    return _status;
   }
 
   // current user
@@ -37,9 +44,19 @@ class AuthServices {
   }
 
   // reset password
-  Future<void> resetPassword(String emailAddress) async {
-    return await _auth.sendPasswordResetEmail(email: emailAddress);
+  Future<AuthResultStatus> resetPassword(String emailAddress) async {
+    try {
+      var results = await _auth.sendPasswordResetEmail(email: emailAddress);
+      _status = AuthResultStatus.successful;
+    } catch(e) {
+      print("*********************>>>>>>>>>>>>>>>>");
+      if (e.toString() == "[firebase_auth/user-not-found] There is no user record corresponding to this identifier. The user may have been deleted."){
+        _status = AuthResultStatus.invalidEmail;
+      }
+    }
+    return _status;
   }
+
 
   // logout
   Future<void> signOut() {